apps/ai-service/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-dotenv
supabase
numpy
opencv-python
deepface
tensorflow
pillow
ultralytics
requests
omegaconf
resend